frontend. factor into own component & mobile responsive layout #94
